@@ -5,18 +5,19 @@ author: Shakti S.P Swain<shakti.swain.mi@gmail.com>
 homepage: https://github.com/Mindinventory/image-collage-widget.git
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  flutter_staggered_grid_view: ^0.3.0
-  flutter_bloc: ^0.21.0
-  flutter_file_manager: ^0.2.0
-  permission_handler: ^3.2.2
-  image_picker: ^0.6.1+4
-  equatable: ^0.6.1
+  flutter_staggered_grid_view: ^0.4.1
+  flutter_bloc: ^8.0.0
+  file_manager: ^1.0.0
+  file_picker: ^4.3.0
+  permission_handler: ^8.3.0
+  image_picker: ^0.8.4+4
+  equatable: ^2.0.3
 
 
 dev_dependencies:
